code review changes, add arg for the buffer size, add arg for the number of threads, add json view
