+++ ds3_java_cli/src/main/java/com/spectralogic/ds3cli/Arguments.java
@@ -48,9 +48,6 @@ public class Arguments {
     private final String[] args;
     private String objectName;
     private String proxy;
-    //TODO re-add these calls when we have support for partial file gets in the helper functions
-    //private int start;
-    //private int end;
     private int retries = 20;
     private Priority priority;
     private WriteOptimization writeOptimization;
@@ -66,6 +63,8 @@ public class Arguments {
 
     private String version = "N/a";
     private String buildDate = "N/a";
+    private String bufferSize;
+    private String numberOfThreads;
 
     public Arguments(final String[] args) throws BadArgumentException, ParseException {
         loadProperties();
@@ -90,11 +89,6 @@ public class Arguments {
         prefix.setArgName("prefix");
         final Option proxy = new Option("x", true, "The URL of the proxy server to use or have \"http_proxy\" set as an environment variable");
         proxy.setArgName("proxy");
-        //TODO re-add these calls when we have support for partial file gets in the helper functions
-        //final Option start = new Option(null, true, "The starting byte for a get_object command");
-        //start.setArgName("start");
-        //final Option end = new Option(null, true, "The ending byte for a get_object command");
-        //end.setArgName("end");
         final Option id = new Option("i", true, "ID for identifying ds3 api resources");
         id.setArgName("id");
         final Option force = new Option(null, false, "Used to force an operation even if there is an error");
@@ -133,6 +127,10 @@ public class Arguments {
         numberOfFiles.setArgName("numberOfFiles");
         final Option sizeOfFiles = new Option("s", true, "The size (in MB) for each file in the performance test");
         sizeOfFiles.setArgName("sizeOfFiles");
+        final Option bufferSize = new Option("bs", true, "Set the buffer size in bytes");
+        bufferSize.setArgName("bufferSize");
+        final Option numberOfThreads = new Option("nt", true, "Set the number of threads");
+        numberOfThreads.setArgName("numberOfThreads");
         options.addOption(ds3Endpoint);
         options.addOption(bucket);
         options.addOption(directory);
@@ -143,9 +141,6 @@ public class Arguments {
         options.addOption(prefix);
         options.addOption(proxy);
         options.addOption(id);
-        //TODO re-add these calls when we have support for partial file gets in the helper functions
-        //options.addOption(start);
-        //options.addOption(end);
         options.addOption(force);
         options.addOption(retries);
         options.addOption(checksum);
@@ -163,6 +158,8 @@ public class Arguments {
         options.addOption(sync);
         options.addOption(numberOfFiles);
         options.addOption(sizeOfFiles);
+        options.addOption(bufferSize);
+        options.addOption(numberOfThreads);
 
         processCommandLine();
     }
@@ -268,17 +265,6 @@ public class Arguments {
         this.setPrefix(cmd.getOptionValue("p"));
         this.setId(cmd.getOptionValue("i"));
 
-        //TODO re-add these calls when we have support for partial file gets in the helper functions
-        //final String start = cmd.getOptionValue();
-        //if (!(start == null || start.isEmpty())) {
-        //    this.setStart(Integer.parseInt(start));
-        //}
-        //final String end = cmd.getOptionValue();
-        //if (!(end == null || end.isEmpty())) {
-        //    this.setEnd(Integer.parseInt(end));
-        //}
-
-
         if (getEndpoint() == null) {
             final String endpoint = System.getenv("DS3_ENDPOINT");
             if (endpoint == null) {
@@ -324,6 +310,16 @@ public class Arguments {
 
         this.setNumberOfFiles(cmd.getOptionValue("n"));
         this.setSizeOfFiles(cmd.getOptionValue("s"));
+
+        this.setBufferSize(cmd.getOptionValue("bs"));
+        if (getBufferSize() == null) {
+            this.bufferSize = "1048576"; //default to 1MB
+        }
+
+        this.setNumberOfThreads(cmd.getOptionValue("nt"));
+        if (getNumberOfThreads() == null) {
+            this.numberOfThreads = "20"; //default to 20 threads
+        }
     }
 
     private WriteOptimization processWriteOptimization(final CommandLine cmd, final String writeOptimization) throws BadArgumentException {
@@ -455,23 +451,6 @@ public class Arguments {
         return this.proxy;
     }
 
-    //TODO re-add these calls when we have support for partial file gets in the helper functions
-    //public int getStart() {
-    //    return start;
-    //}
-
-    //void setStart(final int start) {
-    //    this.start = start;
-    //}
-
-    //public int getEnd() {
-    //    return end;
-    //}
-
-    //void setEnd(final int end) {
-    //    this.end = end;
-    //}
-
     void setForce(final boolean force) {
         this.force = force;
     }
@@ -582,5 +561,21 @@ public class Arguments {
 
     public String getSizeOfFiles() {
         return this.sizeOfFiles;
+    }
+
+    void setBufferSize(final String bufferSize) {
+        this.bufferSize = bufferSize;
+    }
+
+    public String getBufferSize() {
+        return this.bufferSize;
+    }
+
+    void setNumberOfThreads(final String numberOfThreads) {
+        this.numberOfThreads = numberOfThreads;
+    }
+
+    public String getNumberOfThreads() {
+        return this.numberOfThreads;
     }
 }
